Updates to UI appearance, greeting, and overview uses HealthData
