+++ Frontend/composeApp/src/commonMain/kotlin/com/teamnotfound/airise/data/serializable/HealthData.kt
@@ -1,0 +1,10 @@
+package com.teamnotfound.airise.data.serializable
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class HealthData(
+    val caloriesBurned: Int? = 0,
+    val steps: Int? = 0,
+    val avgHeartRate: Int? = 0
+)
