+++ Frontend/composeApp/src/commonMain/kotlin/com/teamnotfound/airise/generativeAi/GeminiApi.kt
@@ -1,6 +1,7 @@
 package com.teamnotfound.airise.generativeAi
 
 import com.teamnotfound.airise.BuildKonfig
+import com.teamnotfound.airise.data.serializable.HealthData
 import dev.shreyaspatil.ai.client.generativeai.Chat
 import dev.shreyaspatil.ai.client.generativeai.GenerativeModel
 import dev.shreyaspatil.ai.client.generativeai.type.Content
@@ -8,9 +9,10 @@ import dev.shreyaspatil.ai.client.generativeai.type.GenerateContentResponse
 import dev.shreyaspatil.ai.client.generativeai.type.PlatformImage
 import dev.shreyaspatil.ai.client.generativeai.type.content
 import kotlinx.coroutines.flow.Flow
+import kotlinx.serialization.json.Json
 
 class GeminiApi {
-    private val PROMPT_TODAYS_OVERVIEW = "With a 100 word limit and , write a fitness summary for today to a user using fake data from the perspective a of a coach."
+    private val promptTodaysOverview = "With a 100 word limit and , write a fitness summary for today to a user using the data provided from the perspective a of a coach."
     private val apiKey = BuildKonfig.GEMINI_API_KEY
 
 
@@ -27,8 +29,10 @@ class GeminiApi {
     suspend fun generateContent(prompt: String): GenerateContentResponse {
         return generativeModel.generateContent(prompt)
     }
-    suspend fun generateTodaysOverview(): GenerateContentResponse {
-        return  generativeModel.generateContent(PROMPT_TODAYS_OVERVIEW)
+    suspend fun generateTodaysOverview(healthData: HealthData): GenerateContentResponse {
+        val healthDataString = Json.encodeToString(healthData)
+        val promptWithData = "Use the following health data for today's overview:\n$healthDataString\n\n$promptTodaysOverview"
+        return generativeModel.generateContent(promptWithData)
     }
 
     fun generateContent(prompt: String, imageData: ByteArray): Flow<GenerateContentResponse> {
